// src/types/NewOpportunityType.ts
export type NewOpportunityType = {
  id: string;
  organizerType: string;
  organizerName: string;
  phone: string;
  link: string;
  eventLink: string;
  organizerPhoto: string;
  title: string;
  categoryName: string;
  opportunityType: string;
  assistanceType: string;
  target: string | number;
  region: string;
  address: string;
  startDate: string;
  endDate: string;
  startHour: string;
  startMinute: string;
  endHour: string;
  endMinute: string;
  startPeriod: string;
  endPeriod: string;
  timeDemands: string;
  skills: string;
  coverImage: string;
  description: string;
  descriptionLink: string;
  document: string;
  currentProgress: string | number;
  status: string;
};